requirements/dev.txt
-r prod.txt

# Static code metrics
# pyflakes==2.0.0
# Flake8==3.6.0
pylint==2.1.1
Radon==2.4.0
coverage==4.5.1

# Testing
pytest==3.9.2
pytest-runner==4.2
pytest-cov==2.6.0
pytest-bdd==3.0.1
pytest-xprocess==0.12.1